전면수정
