+++ django_app/member/models/__init__.py
@@ -1,3 +1,3 @@
-from .myuser import MyUser, Enrollment
-from .tutor import Tutor
-from .certification import Certification
+from .myuser import *
+from .tutor import *
+from .certification import *
