+++ django_app/regiclass/models/__init__.py
@@ -1,4 +1,5 @@
-from .lecture import Lecture, LikeLecture, Enrollment
+from .lecture import *
 from .classlocation import ClassLocation
 from .lecturephoto import LecturePhoto
 from .review import Review
+from .enrollment import Enrollment
