+++ django_app/regiclass/serializers/enrollment.py
@@ -5,31 +5,16 @@ from member.models import MyUser, Tutor
 from regiclass.models import Lecture, Enrollment
 
 
-class MyUserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = MyUser
-        fields = (
-            'username'
-        )
-
-
-class ProcessGuideSerializer(serializers.ModelSerializer):
-    username = MyUserSerializer()
-
-    class Meta:
-        model = Tutor
-        fields = (
-            'username',
-            'author.nickname',
-        )
+class ProcessGuideSerializer(serializers.Serializer):
+    username = serializers.CharField()
+    tutor_nickname = serializers.CharField()
 
 
 class CheckLocationSerializer(serializers.ModelSerializer):
     class Meta:
         model = Enrollment
         fields = (
-            # 'location2',
-            # 'class_weekday',
-            # 'class_time',
+            'location',
+
 
         )
